requirements.txt
playwright>=1.40.0
transformers>=4.57.0
torch>=2.0.0
rich>=13.0.0
typer>=0.14.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
browser-cookie3>=0.20.0